Extract processor contracts
