+++ packages/contracts/source/contracts/consensus/index.ts
@@ -1,2 +1,3 @@
 export * from "./consensus";
 export * from "./enums";
+export * from "./processor";
