+++ packages/contracts/source/contracts/consensus/processor.ts
@@ -1,0 +1,18 @@
+import { ICommittedBlock, IPrecommit, IPrevote, IProposal } from "../crypto";
+import { ProcessorResult } from "./enums";
+
+export interface IProposalProcessor {
+	process(proposal: IProposal, broadcast?: boolean): Promise<ProcessorResult>;
+}
+
+export interface IPrevoteProcessor {
+	process(prevote: IPrevote, broadcast?: boolean): Promise<ProcessorResult>;
+}
+
+export interface IPrecommitProcessor {
+	process(prevote: IPrecommit, broadcast?: boolean): Promise<ProcessorResult>;
+}
+
+export interface ICommittedBlockProcessor {
+	process(committedBlock: ICommittedBlock, broadcast?: boolean): Promise<ProcessorResult>;
+}
